Re-arranging functions, cleanups and comments
